# perf: `offsetToNeedle` caching