--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ coverage
 istanbul-coverage
 fixture-coverage
 test/fixtures/**/dist/
+bench-profile