--- a/eslint.config.js
+++ b/eslint.config.js
@@ -70,7 +70,7 @@ export default defineConfig([
     },
   },
   {
-    ignores: ["dist/**", "test/fixtures/**"],
+    ignores: ["dist/**", "test/fixtures/**", ...tsconfig.exclude],
   },
   eslintPluginPrettierRecommended,
 ]);