--- a/test/location.test.ts
+++ b/test/location.test.ts
@@ -1,5 +1,6 @@
+import { TraceMap } from "@jridgewell/trace-mapping";
 import { test, expect } from "vitest";
-import { offsetToNeedle } from "../src/location";
+import { createEmptySourceMap, Locator } from "../src/location";
 
 const code = `\
 function sum(a, b) {
@@ -19,35 +20,38 @@ function remainder(a, b) {
 }
 Object.defineProperty(__vite_ssr_exports__, "remainder", { enumerable: true, configurable: true, get(){ return remainder }});
 `;
+const map = new TraceMap(createEmptySourceMap("example.js", code));
 
 test("offsetToNeedle picks correct positions", () => {
+  const locator = new Locator(code, map);
+
   expect({
-    start: offsetToNeedle(0, code),
-    end: offsetToNeedle(38, code),
+    start: locator.offsetToNeedle(0),
+    end: locator.offsetToNeedle(38),
   }).toEqual({
     start: { line: 1, column: 0 },
     end: { line: 3, column: 1 },
   });
 
   expect({
-    start: offsetToNeedle(153, code),
-    end: offsetToNeedle(196, code),
+    start: locator.offsetToNeedle(153),
+    end: locator.offsetToNeedle(196),
   }).toEqual({
     start: { line: 5, column: 0 },
     end: { line: 7, column: 1 },
   });
 
   expect({
-    start: offsetToNeedle(321, code),
-    end: offsetToNeedle(364, code),
+    start: locator.offsetToNeedle(321),
+    end: locator.offsetToNeedle(364),
   }).toEqual({
     start: { line: 9, column: 0 },
     end: { line: 11, column: 1 },
   });
 
   expect({
-    start: offsetToNeedle(489, code),
-    end: offsetToNeedle(533, code),
+    start: locator.offsetToNeedle(489),
+    end: locator.offsetToNeedle(533),
   }).toEqual({
     start: { line: 13, column: 0 },
     end: { line: 15, column: 1 },
@@ -61,10 +65,14 @@ function sum(a, b) {
   return a + b;
 }
   `;
+  const locator = new Locator(
+    code,
+    new TraceMap(createEmptySourceMap("example.js", code)),
+  );
 
   expect({
-    start: offsetToNeedle(58, code),
-    end: offsetToNeedle(61, code),
+    start: locator.offsetToNeedle(58),
+    end: locator.offsetToNeedle(61),
   }).toEqual({
     start: { line: 2, column: 9 },
     end: { line: 2, column: 12 },
@@ -81,6 +89,11 @@ function sum(a, b) {
     .split("\n")
     .join("\r\n");
 
-  expect(offsetToNeedle(58, code)).toEqual({ line: 2, column: 9 });
-  expect(offsetToNeedle(61, code)).toEqual({ line: 2, column: 12 });
+  const locator = new Locator(
+    code,
+    new TraceMap(createEmptySourceMap("example.js", code)),
+  );
+
+  expect(locator.offsetToNeedle(58)).toEqual({ line: 2, column: 9 });
+  expect(locator.offsetToNeedle(61)).toEqual({ line: 2, column: 12 });
 });