--- a/tsconfig.json
+++ b/tsconfig.json
@@ -6,5 +6,6 @@
     "moduleResolution": "bundler",
     "types": ["./test/utils/expect-extend"]
   },
-  "include": ["src/*", "test/*"]
+  "include": ["src/*", "test/*", "benchmark/*"],
+  "exclude": ["benchmark/fixtures/*"]
 }