… benchmark/fixtures/checker.ts
… benchmark/fixtures/vitest-cli-api-bundled.js