- `checker.ts` conversion goes from 1210s to 34s
- `vitest-cli-api-bundled.js` goes from 44s to 1.7s

Before-after:

<img src="https://github.com/user-attachments/assets/7a75aeb0-796c-4424-a473-48cd3295efce" width="400" />

<img src="https://github.com/user-attachments/assets/548d44d4-a39c-4aad-9f24-ab940651436e" width="400" />